% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publication.R
\name{add_model,Publication,FixedEffectsModel-method}
\alias{add_model,Publication,FixedEffectsModel-method}
\title{Add a model to a publication}
\usage{
\S4method{add_model}{Publication,FixedEffectsModel}(publication, model)
}
\arguments{
\item{publication}{The publication for which a set will be added}

\item{model}{The model to add to the publication}
}
\value{
A publication with the added model
}
\description{
This function adds objects of class \code{FixedEffectsModel} or
\code{MixedEffectsModel} to a publication. Models added in this way are added as
a set containing only one model. This operation is not done in-place.
}
\keyword{internal}
