% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{add_model}
\alias{add_model}
\title{Add a model to a publication}
\usage{
add_model(publication, model)
}
\arguments{
\item{publication}{The publication for which a set will be added.}

\item{model}{The model to add to the publication}
}
\description{
This function adds objects of class \code{FixedEffectsModel} or
\code{MixedEffectsModel} to a publication. Models added in this way are added as
a set containing only one model.
}
