% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_models.R
\name{aggregate_pub_models}
\alias{aggregate_pub_models}
\title{Aggregates the set of models in a publication into a model_tbl}
\usage{
aggregate_pub_models(pub, current_ids = c())
}
\arguments{
\item{pub}{The publication object}

\item{current_ids}{An optional vector of currently ingestedmodel  IDs.}
}
\description{
Aggregates the set of models in a publication into a model_tbl
}
\keyword{internal}
