% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_models.R
\name{aggregate_results}
\alias{aggregate_results}
\title{Transforms a set of searched models into a tibble of models and descriptors}
\usage{
aggregate_results(results)
}
\description{
Transforms a set of searched models into a tibble of models and descriptors
}
\keyword{internal}
