% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_models.R
\name{aggregate_results_ext}
\alias{aggregate_results_ext}
\title{Transforms a set of searched models into a tibble of models and descriptors}
\usage{
aggregate_results_ext(results)
}
\description{
Transforms a set of searched models into a tibble of models and descriptors
}
