% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometric.R
\docType{package}
\name{allometric-package}
\alias{allometric}
\alias{allometric-package}
\title{allometric: Structured Allometric Models for Trees}
\description{
To learn more about allometric, refer to the
\href{https://allometric.org}{documentation website}.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/allometric/allometric/issues}
}

}
\author{
\strong{Maintainer}: Bryce Frank \email{bfrank70@gmail.com} (\href{https://orcid.org/0000-0003-2980-6860}{ORCID})

Authors:
\itemize{
  \item Bryce Frank \email{bfrank70@gmail.com} (\href{https://orcid.org/0000-0003-2980-6860}{ORCID})
  \item Francisco Mauro \email{francisco.mauro@uva.es} (\href{https://orcid.org/0000-0002-7832-6268}{ORCID})
  \item Elijah Allensworth \email{elijah.allensworth@protonmail.com}
}

}
