% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricModel.R
\name{AllometricModel}
\alias{AllometricModel}
\title{Base class for allometric models}
\usage{
AllometricModel(
  response_unit,
  covariate_units,
  predict_fn,
  descriptors = list(),
  covariate_definitions = list()
)
}
\arguments{
\item{response_unit}{A named list containing one element, with a name representing the response
variable and a value representing the units of the response variable
using the \code{units::as_units} function.}

\item{covariate_units}{A named list containing the covariate specifications, with names
representing the covariate name and the values representing the units of
the coavariate using the \code{units::as_units} function.}

\item{descriptors}{An optional list of descriptors that are specified at the model-level.}

\item{covariate_definitions}{A named list of covariate definitions that
are used instead of the descriptions given by the variable naming system.}
}
\description{
This class is primarily used as a parent class for other model
implementations.
}
\keyword{internal}
