% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{as_character_or_null}
\alias{as_character_or_null}
\title{Coerce a filter argument to a character vector, or NULL if empty}
\usage{
as_character_or_null(x)
}
\description{
Coerce a filter argument to a character vector, or NULL if empty
}
\keyword{internal}
