\name{bell_1981}
\alias{bell_1981}
\title{Bell et al. (1981)}
\description{Allometric models from Bell et al. (1981)}
\section{Stem Volume Models}{\code{vsa = 10^a * dsob^b * hst^c}
\itemize{
\item{\code{vsa [ft3]}}{ - volume of the entire stem, including top and stump}
\item{\code{dsob [in]}}{ - diameter of the stem, outside bark at breast height}
\item{\code{hst [ft]}}{ - total height of the stem}
}
\bold{Model Parameters}
\preformatted{      a     b     c
1 -2.96  1.81  1.27}
\out{<hr>}}
