% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brackett_acer}
\alias{brackett_acer}
\title{Brackett Acer Volume Model}
\format{
An object of class \code{FixedEffectsModel} of length 1.
}
\usage{
brackett_acer
}
\description{
An example allometric model that predicts volume for the genus \emph{Acer}.
}
\keyword{datasets}
