% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brackett_rubra}
\alias{brackett_rubra}
\title{An object of class \code{FixedEffectsModel}}
\format{
An object of class \code{FixedEffectsModel} of length 1.
}
\usage{
brackett_rubra
}
\description{
Brackett Rubra Volume Model
}
\details{
An example allometric model that predicts volume for \emph{Alnus rubra}.
}
\keyword{datasets}
