% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_args_in_predict_fn}
\alias{check_args_in_predict_fn}
\title{Are the arguments given in the predict function inside the body?}
\usage{
check_args_in_predict_fn(object)
}
\description{
Are the arguments given in the predict function inside the body?
}
