% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_checks.R
\name{check_citation_key}
\alias{check_citation_key}
\title{Checks if a citation contains a key}
\usage{
check_citation_key(citation)
}
\description{
Checks if a citation contains a key
}
\keyword{internal}
