% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_checks.R
\name{check_country_in_iso}
\alias{check_country_in_iso}
\title{Checks if a country code is defined in ISO_3166_1$Alpha_2}
\usage{
check_country_in_iso(country_codes)
}
\description{
Checks if a country code is defined in ISO_3166_1$Alpha_2
}
\keyword{internal}
