% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{check_covariates_equal}
\alias{check_covariates_equal}
\title{Check for equivalence of covariate slots in two models}
\usage{
check_covariates_equal(mod1, mod2)
}
\arguments{
\item{mod1}{A model object}

\item{mod2}{A model object}
}
\description{
The covariate slots are considered equal if they contain the same number of
covariates, the names match, the units match, and the covariates are in the
same order.
}
\keyword{internal}
