% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_checks.R
\name{check_covts_in_args}
\alias{check_covts_in_args}
\title{Check if all covariates in covariate_units are used as arguments in
predict_fn}
\usage{
check_covts_in_args(object)
}
\description{
Check if all covariates in covariate_units are used as arguments in
predict_fn
}
\keyword{internal}
