% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publication.R
\name{check_descriptor_set}
\alias{check_descriptor_set}
\title{Check validity of descriptors}
\usage{
check_descriptor_set(descriptor_set)
}
\description{
There must not be any duplicate names across the three sets
(pub, set, model). If there is, stop execution with error that names the
duplicated descriptors
}
\keyword{internal}
