% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_checks.R
\name{check_descriptor_validity}
\alias{check_descriptor_validity}
\title{Checks that descriptors are valid}
\usage{
check_descriptor_validity(descriptors)
}
\description{
Checks that descriptors are valid
}
\keyword{internal}
