% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{check_filter_values}
\alias{check_filter_values}
\title{Validate a filter argument against the values present in the data}
\usage{
check_filter_values(value, arg, available)
}
\description{
Validate a filter argument against the values present in the data
}
\keyword{internal}
