% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{check_ids_equal}
\alias{check_ids_equal}
\title{Check for equivalence of ID slots in two models}
\usage{
check_ids_equal(mod1, mod2)
}
\arguments{
\item{mod1}{A model object}

\item{mod2}{A model object}
}
\value{
TRUE if equal, FALSE if not
}
\description{
Two models are considered to have equal IDs if both IDs are unset (NA), or
both IDs are set and equal. They are not equal if one is set and the other
is not, or if both are set and differ.
}
\keyword{internal}
