% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_local_models}
\alias{check_local_models}
\title{Check allometric models can be installed from local files}
\usage{
check_local_models(publications_path, parameters_path, verbose = TRUE)
}
\arguments{
\item{publications_path}{The path to publication files}

\item{parameters_path}{The path to parameter files}

\item{verbose}{Whether or not to print verbose messages}
}
\description{
For the purposes of contributing models, it is useful for contributors to
run a local set of allometric model files. For example, when contributors
add new publication files to \code{allometric/models}, this function can be used
to ensure that the local model files run correctly.
}
