% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_models_installed}
\alias{check_models_installed}
\title{Check if allometric models are currently installed}
\usage{
check_models_installed(verbose = FALSE)
}
\arguments{
\item{verbose}{Print verbose messages if TRUE}
}
\description{
Check if allometric models are currently installed
}
