% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_checks.R
\name{check_parameters_in_mixed_fns}
\alias{check_parameters_in_mixed_fns}
\title{Check if all parameter_names are in predict_fn and predict_ranef}
\usage{
check_parameters_in_mixed_fns(object)
}
\description{
Check if all parameter_names are in predict_fn and predict_ranef
}
\keyword{internal}
