% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_checks.R
\name{check_region_in_iso}
\alias{check_region_in_iso}
\title{Checks if a region code is defined in the ISO_3166_2 table}
\usage{
check_region_in_iso(region_code)
}
\description{
Checks if a region code is defined in the ISO_3166_2 table
}
\keyword{internal}
