% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{check_response_equal}
\alias{check_response_equal}
\title{Check for equivalence of response slots in two models}
\usage{
check_response_equal(mod1, mod2)
}
\arguments{
\item{mod1}{A model object}

\item{mod2}{A model object}
}
\description{
The response slots are considered equal if the names match and the units
match. Units are first parsed to strings and then checked for equivalence
}
\keyword{internal}
