% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_models.R
\name{check_run_pub_list}
\alias{check_run_pub_list}
\title{Checks if the pub_list should be generated}
\usage{
check_run_pub_list(pub_list_path)
}
\description{
The pub_list is regenerated if any file in inst/publications has been
modified after the creation of the pub_list
}
\keyword{internal}
