% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa.R
\name{check_taxa_unique}
\alias{check_taxa_unique}
\title{Determines if a taxa is composed of unique taxon objects}
\usage{
check_taxa_unique(object)
}
\description{
Determines if a taxa is composed of unique taxon objects
}
\keyword{internal}
