% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_checks.R
\name{check_taxon_hierarchy}
\alias{check_taxon_hierarchy}
\title{Determines if a taxon has a valid hierarchy}
\usage{
check_taxon_hierarchy(object)
}
\description{
Determines if a taxon has a valid hierarchy
}
\keyword{internal}
