% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_defs.R
\name{check_valid_search_str}
\alias{check_valid_search_str}
\title{Check if the search string is valid}
\usage{
check_valid_search_str(num_underscores)
}
\description{
Search strings can have prefixes and suffixes, but these must only be one
character.
}
\keyword{internal}
