% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{covariate_unit_strs}
\alias{covariate_unit_strs}
\title{Units of a covariate list, as strings ("" for unitless)}
\usage{
covariate_unit_strs(covariates)
}
\description{
Units of a covariate list, as strings ("" for unitless)
}
\keyword{internal}
