% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tbl.R
\name{create_model_row}
\alias{create_model_row}
\title{Creates a dataframe row from model information}
\usage{
create_model_row(model)
}
\description{
Creates a dataframe row from model information
}
\keyword{internal}
