% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{descriptors<-}
\alias{descriptors<-}
\title{Set the descriptors of a model.}
\usage{
descriptors(model) <- value
}
\arguments{
\item{value}{A tibble::tbl_df of descriptors}
}
\description{
Set the descriptors of a model.
}
\keyword{internal}
