% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{descriptors_json_to_list}
\alias{descriptors_json_to_list}
\title{Parse a JSON descriptor column to a named list}
\usage{
descriptors_json_to_list(json)
}
\description{
Parse a JSON descriptor column to a named list
}
\keyword{internal}
