% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromJSON.R
\name{descriptors_to_S4}
\alias{descriptors_to_S4}
\title{Convert the descriptors JSON data to a named list of descriptors}
\usage{
descriptors_to_S4(descriptors_data)
}
\description{
Convert the descriptors JSON data to a named list of descriptors
}
\keyword{internal}
