% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffectsModel.R
\name{==,FixedEffectsModel,FixedEffectsModel-method}
\alias{==,FixedEffectsModel,FixedEffectsModel-method}
\title{Check equivalence of fixed effects models}
\usage{
\S4method{==}{FixedEffectsModel,FixedEffectsModel}(e1, e2)
}
\arguments{
\item{e1}{A \code{FixedEffectsModel} object}

\item{e2}{A \code{FixedEffectsModel} object}
}
\description{
Fixed effects models are considered equal if all of the following are true:
\itemize{
\item{The model IDs are equal (or not present)}
\item{The response unit names and units are the same}
\item{The covariate names and units are the same and are in the same order}
\item{The specification names and values are the same}
\item{The predict_fn is the same}
\item{The response definitions are the same}
\item{The covariate definitions are the same}
}
}
