% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedEffectsModel.R
\name{==,MixedEffectsModel,MixedEffectsModel-method}
\alias{==,MixedEffectsModel,MixedEffectsModel-method}
\title{Check equivalence of mixed effects models}
\usage{
\S4method{==}{MixedEffectsModel,MixedEffectsModel}(e1, e2)
}
\arguments{
\item{e1}{A \code{MixedEffectsModel} object}

\item{e2}{A \code{MixedEffectsModel} object}
}
\description{
Fixed effects models are considered equal if all of the following are true:
\itemize{
\item{The model IDs are equal (or not present)}
\item{The response unit names and units are the same}
\item{The covariate names and units are the same and are in the same order}
\item{The specification names and values are the same}
\item{The predict_fn are the same}
\item{The predict_ranef are the same}
\item{The fixed_only slots are the same}
\item{The response definitions are the same}
\item{The covariate definitions are the same}
}
}
