% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxon.R
\name{==,Taxon,Taxon-method}
\alias{==,Taxon,Taxon-method}
\title{Check equivalence of two \code{Taxon} objects}
\usage{
\S4method{==}{Taxon,Taxon}(e1, e2)
}
\arguments{
\item{e1}{A \code{Taxon} object}
}
\value{
TRUE or FALSE
}
\description{
Two Taxons are equal if all values are the same (including NA values)
}
\keyword{internal}
