% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet.R
\name{family_names_from_author}
\alias{family_names_from_author}
\title{Extract author family names from a BibTeX author string}
\usage{
family_names_from_author(author)
}
\description{
Extract author family names from a BibTeX author string
}
\keyword{internal}
