% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_defs.R
\name{get_component_defs}
\alias{get_component_defs}
\title{Load the component definitions}
\usage{
get_component_defs()
}
\value{
A tibble::tbl_df containing the component definitions
}
\description{
Loads the component definitions from a locally stored csv file
}
\examples{
get_component_defs()
}
