% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_defs.R
\name{get_measure_defs}
\alias{get_measure_defs}
\title{Load the measure definitions}
\usage{
get_measure_defs()
}
\value{
A tibble::tbl_df containing the measure definitions
}
\description{
Loads the measure definitions from a locally stored csv file
}
