% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_model}
\alias{get_model}
\title{Get a model with a known \code{model_id}}
\usage{
get_model(model_id, citation = TRUE)
}
\arguments{
\item{model_id}{The 8 character uuid of a model}
}
\description{
This function uses the \verb{model/<model_id>} endpoint to retrieve a model
with a known \code{model_id}.
}
