% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_processing.R
\name{get_model_hash}
\alias{get_model_hash}
\title{Hashes a function string}
\usage{
get_model_hash(predict_fn_populated, descriptors)
}
\description{
We need some sort of stable data structure that will serve as a unique ID
for a model, but will also change in the event that the model changes. This
way, models can be "versioned" across time, which may be useful for debugging
purposes down the line. This function trims whitespace and lowercases
the predict_fn_populated pasted with the descriptors, which serves as a
reasonable proxy for the model.
}
\keyword{internal}
