% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_models.R
\name{get_model_proxy_id}
\alias{get_model_proxy_id}
\title{Concats the authors, title and year and model index to create a proxy ID
string for a model.}
\usage{
get_model_proxy_id(
  last_names,
  title,
  year,
  response_set_ix,
  model_set_ix,
  model_ix
)
}
\description{
Concats the authors, title and year and model index to create a proxy ID
string for a model.
}
