% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricModel.R
\name{get_model_type}
\alias{get_model_type}
\title{Gets the model type for a response name.}
\usage{
get_model_type(response_name)
}
\arguments{
\item{response_name}{The response_name from the variable naming system.}
}
\description{
From the variable naming system, return the model type. Model types are
custom names that are easier to understand than the usual component-measure
pairing. For example, site index models would be called "stem height models"
hecause the component is the stem and the measure is the height. However,
site index models need a special name. Model types are these names. If a
model type is not defined, the component-measure pairing is used instead.
}
\keyword{internal}
