% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_params_path}
\alias{get_params_path}
\title{Get the parameter search path}
\usage{
get_params_path()
}
\value{
A string containing the currently set parameter search path
}
\description{
Get the parameter search path
}
