% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_models.R
\name{get_pub_file_spec}
\alias{get_pub_file_spec}
\title{Retrieve all publication file names in the publication subdirectories}
\usage{
get_pub_file_spec(pub_path)
}
\description{
The publication subdirectories divide the publications into groups sorted
by last name. This function returns the publication file paths in alphabetic
order.
}
\keyword{internal}
