% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_processing.R
\name{get_pub_file_specs}
\alias{get_pub_file_specs}
\title{Retrieve all publication file names in the publication subdirectories}
\usage{
get_pub_file_specs(pub_path)
}
\description{
The publication subdirectories divide the publications into groups sorted
by last name. This function returns a list containing the publication paths
and the publication names.
}
\keyword{internal}
