% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa.R
\name{\%in\%,character,Taxa-method}
\alias{\%in\%,character,Taxa-method}
\title{Check if a character is in a Taxa}
\usage{
\S4method{\%in\%}{character,Taxa}(x, table)
}
\arguments{
\item{x}{A character vector}

\item{table}{A \code{Taxa} object}
}
\value{
TRUE or FALSE indicating if the character appears in the Taxa fields
}
\description{
Check if a character is in a Taxa
}
\keyword{internal}
