% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxon.R
\name{\%in\%,character,Taxon-method}
\alias{\%in\%,character,Taxon-method}
\title{Check if a character is in a Taxon}
\usage{
\S4method{\%in\%}{character,Taxon}(x, table)
}
\arguments{
\item{x}{A character vector}

\item{table}{A \code{Taxon} object}
}
\value{
TRUE or FALSE indicating if the character appears in the Taxon fields
}
\description{
Check if a character is in a Taxon
}
\keyword{internal}
