% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa.R
\name{\%in\%,Taxa,character-method}
\alias{\%in\%,Taxa,character-method}
\title{Check if a Taxa contains a character}
\usage{
\S4method{\%in\%}{Taxa,character}(x, table)
}
\arguments{
\item{x}{A \code{Taxa} object}

\item{table}{A character vector}
}
\value{
TRUE or FALSE indicating if any of the Taxa fields appear in the
character.
}
\description{
Check if a Taxa contains a character
}
\keyword{internal}
