% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa.R
\name{\%in\%,Taxa,Taxon-method}
\alias{\%in\%,Taxa,Taxon-method}
\title{Check if a Taxa is in Taxon}
\usage{
\S4method{\%in\%}{Taxa,Taxon}(x, table)
}
\arguments{
\item{x}{A \code{Taxa} object}

\item{table}{A \code{Taxon} object}
}
\value{
A logical vector indicating which elements of \code{Taxa} are equal to
the \code{Taxon}
}
\description{
Check if a Taxa is in Taxon
}
\keyword{internal}
