% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxon.R
\name{\%in\%,Taxon,character-method}
\alias{\%in\%,Taxon,character-method}
\title{Check if a Taxon contains a character}
\usage{
\S4method{\%in\%}{Taxon,character}(x, table)
}
\arguments{
\item{x}{A \code{Taxon} object}

\item{table}{A character vector}
}
\value{
TRUE or FALSE indicating if any of the Taxa fields appear in the
character.
}
\description{
Check if a Taxon contains a character
}
