% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxon.R
\name{\%in\%,Taxon,Taxa-method}
\alias{\%in\%,Taxon,Taxa-method}
\title{Check if a Taxon is in a Taxa}
\usage{
\S4method{\%in\%}{Taxon,Taxa}(x, table)
}
\arguments{
\item{x}{A \code{Taxon} object}

\item{table}{A \code{Taxa} object}
}
\value{
TRUE or FALSE indicating if the Taxon is in the Taxa
}
\description{
Check if a Taxon is in a Taxa
}
\keyword{internal}
