\name{hann_1978}
\alias{hann_1978}
\title{Hann et al. (1978)}
\description{Allometric models from Hann et al. (1978)}
\section{Stem Volume Models}{\code{vsa = a_0 + a_1 * dsob^2 * hst}
\itemize{
\item{\code{vsa [ft3]}}{ - volume of the entire stem, including top and stump}
\item{\code{dsob [in]}}{ - diameter of the stem, outside bark at breast height}
\item{\code{hst [ft]}}{ - total height of the stem}
}
\bold{Model Parameters}
\preformatted{   family     genus       species     national_forest                               a_0     a_1
 1 Pinaceae   Pinus       strobus     Lincoln, Santa Fe, Carson                  0.161  0.00203
 2 Pinaceae   Picea       engelmanii  Lincoln, Santa Fe, Carson                  0.225  0.00217
 3 Pinaceae   Pinus       jeffreyi    Coconino, Tonto, Lincoln, Santa Fe, Carson 0.237  0.00221
 4 Pinaceae   Pinus       ponderosa   Coconino, Tonto, Lincoln                   0.0811 0.00198
 5 Pinaceae   Pinus       ponderosa   Santa Fe, Carson                           0.0483 0.00205
 6 Pinaceae   Pseudotsuga menziesii   Lincoln                                    0.438  0.00176
 7 Pinaceae   Pseudotsuga menziesii   Santa Fe, Carson                           0.341  0.00192
 8 Salicaceae Populus     tremuloides Santa Fe, Carson                           0.0327 0.00231
 9 Pinaceae   Abies       concolor    Lincoln                                    0.211  0.00184
10 Pinaceae   Abies       concolor    Santa Fe, Carson                           0.158  0.00201}
\out{<hr>}}
