% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_models.R
\name{ingest_models}
\alias{ingest_models}
\title{Create allometric models from parameter and publication files}
\usage{
ingest_models(verbose, pub_path = NULL)
}
\description{
This function ingests models by running each publication R file and
populating a dataframe that contains each model object and some metadata.
The result of this function creates the table of models obtained using
\code{load_models()}. See \code{install_models()} for the end-user entrypoint.
}
\keyword{internal}
