% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_model_rows}
\alias{load_model_rows}
\title{Shared lookup for load_model() and load_set()}
\usage{
load_model_rows(pub_id, name)
}
\description{
Shared lookup for load_model() and load_set()
}
\keyword{internal}
