% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_processing.R
\name{map_publications}
\alias{map_publications}
\title{Iteratively process publication files}
\usage{
map_publications(verbose, func, pub_path = NULL, params_path = NULL)
}
\arguments{
\item{verbose}{Whether or not to print verbose messages to console}

\item{func}{The publication processing function. It should take a Publication
object as its only argument.}

\item{pub_path}{An optional path to a publication directory, by
default the internally stored set of publications is used.}

\item{params_path}{An optional path to a parameters directory, by
default the internally stored set of parameter files is used.}
}
\description{
This function allows a user to flexibly extract information as it loops over
the publication files. Two main internal use-cases exist for this. First,
it is used to install models as is done in \code{insall_models()} and, second,
it is used to populate the remote MongoDB. Most users will not be interested
in this function, but it is exposed for usage in the \code{allodata} package.
}
