% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tbl.R
\name{merge.model_tbl}
\alias{merge.model_tbl}
\title{Merge a \code{model_tbl} with another data frame.}
\usage{
\method{merge}{model_tbl}(x, y, ...)
}
\arguments{
\item{x}{A data frame or \code{model_tbl}}

\item{y}{A data frame or \code{model_tbl}}

\item{...}{Additional arguments passed to \code{merge}}
}
\value{
A \code{model_tbl} merged with the inputs
}
\description{
This merge function ensures that, when \code{model_tbl} is used in a merge that
the resultant dataframe is still a \code{model_tbl}.
}
