% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedEffectsModel.R
\name{model_call,MixedEffectsModel-method}
\alias{model_call,MixedEffectsModel-method}
\title{Get the function call for a model}
\usage{
\S4method{model_call}{MixedEffectsModel}(model)
}
\arguments{
\item{model}{The allometric model for which a function call will be
retrieved}
}
\value{
A string of the function call
}
\description{
The function call is the allometric model expressed as a function of its
covariates. This function allows the user to see the function call.
}
\keyword{internal}
