% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParametricModel.R
\name{model_call,ParametricModel-method}
\alias{model_call,ParametricModel-method}
\title{Get the function call for a model}
\usage{
\S4method{model_call}{ParametricModel}(model)
}
\arguments{
\item{model}{The allometric model for which a function call will be
retrieved}
}
\value{
A string of the function call
}
\description{
The function call is the allometric model expressed as a function of its
covariates. Accessing the function call is important when determining the
order of the covariates given to the prediction function.
}
\examples{
model <- select_model(allometric_models, "f21028ef")
model_call(model)
}
\keyword{internal}
