% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{model_call}
\alias{model_call}
\title{Get the function call for a model}
\usage{
model_call(object)
}
\arguments{
\item{model}{The allometric model for which a function call will be
retrieved}
}
\value{
A string of the function call
}
\description{
The function call is the allometric model expressed as a function of its
covariates. Accessing the function call is important when determining the
order of the covariates given to the prediction function.
}
\examples{
model_call(brackett_rubra)
}
