% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParametricModel.R
\name{ParametricModel}
\alias{ParametricModel}
\title{Base class for all parametric models.}
\usage{
ParametricModel(
  response,
  covariates,
  predict_fn,
  parameters,
  descriptors = list(),
  response_definition = NA_character_,
  covariate_definitions = list()
)
}
\arguments{
\item{response}{A named list containing one element, with a name representing the response
variable and a value representing the units of the response variable
using the \code{units::as_units} function.}

\item{covariates}{A named list containing the covariate specifications, with names
representing the covariate name and the values representing the units of
the coavariate using the \code{units::as_units} function}

\item{predict_fn}{A function that takes the covariate names as arguments and returns a
prediction of the response variable. This function should be vectorized.}

\item{descriptors}{An optional named list of descriptors that describe the context of the
allometric model}

\item{response_definition}{A string containing an optional custom response definition, which is used
instead of the description given by the variable naming system.}

\item{covariate_definitions}{An optional named list of custom covariate definitions that will supersede
the definitions given by the variable naming system. The names of the list
must match the covariate names given in \code{covariates}.}
}
\value{
An object of class ParametricModel
}
\description{
This is a base class used for \code{FixedEffectsModel} and \code{MixedEffectsModel}
}
\keyword{internal}
