% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tbl.R
\name{select_model.model_tbl}
\alias{select_model.model_tbl}
\title{Select a model from \code{allometric_models}}
\usage{
\method{select_model}{model_tbl}(model_tbl, id)
}
\arguments{
\item{model_tbl}{A \code{model_tbl} object}

\item{id}{The model id or index}
}
\value{
An allometric model object
}
\description{
This function is used to select a single model from a \code{model_tbl}
dataframe using its id.
}
\keyword{internal}
