% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tbl.R
\name{select_model}
\alias{select_model}
\title{Select an allometric model}
\usage{
select_model(model_tbl, id)
}
\arguments{
\item{model_tbl}{A \code{model_tbl} object}

\item{id}{The model id or index}
}
\description{
This is a generic function used to select allometric models out of larger
collections, like \code{model_tbl}.
}
