% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{set_params_path}
\alias{set_params_path}
\title{Set the parameter search path}
\usage{
set_params_path(params_path)
}
\arguments{
\item{params_path}{The file path containing parameter files}
}
\description{
Set the parameter search path
}
