% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxa.R
\name{Taxa}
\alias{Taxa}
\title{Group taxons together}
\usage{
Taxa(...)
}
\arguments{
\item{...}{A set of \code{Taxon} objects.}
}
\value{
An instance of class \code{Taxa}
}
\description{
\code{Taxa} represents a set of taxons. See \code{Taxon()}. These are typically used
to specify species and other taxonomic groups that belong to a model.
}
\examples{
Taxa(
   Taxon(
      family = "Pinaceae",
      genus = "Pinus",
      species = "ponderosa"
   ),
   Taxon(
     family = "Betulaceae"
   )
)
}
