% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffectsModel.R
\name{toJSON,FixedEffectsModel-method}
\alias{toJSON,FixedEffectsModel-method}
\title{Convert a fixed effects model to a JSON representation}
\usage{
\S4method{toJSON}{FixedEffectsModel}(object, ...)
}
\arguments{
\item{object}{A fixed effects model}

\item{...}{Additional arguments passed to jsonlite::toJSON}
}
\value{
A string containing the JSON representation of the object
}
\description{
This function converts a fixed effects model into a JSON representation.
Primarily, this is used internally to populate a remotely hosted
MongoDatabase.
}
\examples{
toJSON(brackett_rubra)
}
