% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{toJSON}
\alias{toJSON}
\title{Convert a model or publication to a JSON representation}
\usage{
toJSON(object, ...)
}
\arguments{
\item{object}{An allometric model or publication}

\item{...}{Additional arguments passed to jsonlite::toJSON}
}
\value{
A string containing the JSON representation of the object
}
\description{
This function converts an allometric model or publication into a JSON
representation. Primarily, this is used internally to populate a remotely
hosted MongoDatabase.
}
\examples{
toJSON(brackett_rubra)
}
