\name{turner_2000}
\alias{turner_2000}
\title{Turner et al. (2000)}
\description{Allometric models from Turner et al. (2000)}
\section{Foliage Biomass Models}{\code{bf = exp(a + (b * log(dsob)))}
\itemize{
\item{\code{bf [kg]}}{ - foliage biomass}
\item{\code{dsob [cm]}}{ - diameter of the stem, outside bark at breast height}
}
\bold{Model Parameters}
\preformatted{  family       genus       species          a     b
1 Pinaceae     Pseudotsuga menziesii    -2.85  1.70
2 Pinaceae     Tsuga       heterophylla -4.13  2.13
3 Cupressaceae Thuja       plicata      -2.62  1.78}
\out{<hr>}}
