% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tbl.R
\name{unnest_models.model_tbl}
\alias{unnest_models.model_tbl}
\title{Unnest the columns of \code{model_tbl}}
\usage{
\method{unnest_models}{model_tbl}(data, cols)
}
\arguments{
\item{data}{A \code{model_tbl}}

\item{cols}{A character vector of columns to unnest}
}
\value{
The unnested \code{model_tbl}
}
\description{
A \code{model_tbl} often contains nested information within the cells of the
table. This function allows a user to unnest the columns of interest.
}
\keyword{internal}
