% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tbl.R
\name{unnest_models}
\alias{unnest_models}
\title{Unnest columns of a dataframe}
\usage{
unnest_models(data, cols)
}
\arguments{
\item{data}{A dataframe}

\item{cols}{A character vector indicating the columns to unnest}
}
\value{
The unnested \code{model_tbl}
}
\description{
Unnest columns of a dataframe
}
