% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tbl.R
\name{unnest_taxa}
\alias{unnest_taxa}
\title{Unnest the taxa column of a \code{model_tbl}}
\usage{
unnest_taxa(data)
}
\arguments{
\item{data}{A \code{model_tbl}}
}
\value{
A \code{model_tbl} with family, genus and species columns attached
}
\description{
In some cases it is convenient to expand the taxonomic specifications for
each model contained in the \code{taxa} column. This function achieves this,
and adds \code{family}, \code{genus}, and \code{species} character columns. Models with
more than one taxon are replicated as new rows.
}
