% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{write_model_cache}
\alias{write_model_cache}
\title{Write one filter variant into the model cache}
\usage{
write_model_cache(cache_path, dist_key, filter_key, models, variants)
}
\description{
Write one filter variant into the model cache
}
\keyword{internal}
