% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_processing.R
\name{aggregate_pub_models}
\alias{aggregate_pub_models}
\title{Aggregates the set of models in a publication into a model_tbl}
\usage{
aggregate_pub_models(pub)
}
\arguments{
\item{pub}{The publication object}
}
\description{
Aggregates the set of models in a publication into a model_tbl
}
\keyword{internal}
