% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{aggregate_taxa}
\alias{aggregate_taxa}
\title{Aggregate family, genus, and species columns of `tbl_df`` into taxa data
structure}
\usage{
aggregate_taxa(table, grouping_col = NULL)
}
\arguments{
\item{table}{The table for which the taxa will be aggregated}

\item{grouping_col}{An optional column to group on when creating taxa. Rows
with the same grouping_col value will be stored into the same taxa.}
}
\value{
A tibble with family, genus, and species columns added
}
\description{
This function facilitates aggregating family, genus, and species columns
into the taxa data structure, which is a nested list composed of multiple
"taxons". A taxon is a list containing family, genus, and species fields.
}
