% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{allometric_models}
\alias{allometric_models}
\title{A table of installed allometric models}
\description{
\code{allometric_models}is a \code{tibble::tbl_df} containing every installed
allometric model in \code{allometric}. It is available globally on package load.
If not, run the \code{install_models} function which will install the models and
expose the \code{allometric_models} table in your session.
}
\details{
This table is composed primarily of
\href{https://tidyr.tidyverse.org/articles/nest.html}{nested data}, referred to
as \code{list} columns. Nested data, while somewhat cumbersome for users that are
not experienced with it, is necessary for accommodating the flexible nature
of published allometric models. An obvious motivation is that a given model
contains multiple authors, covariates, even multiple regions or countries.
Nested data allows for the storage of these highly variable structures for
each model.
}
