% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllometricModel.R
\name{AllometricModel}
\alias{AllometricModel}
\title{Base class for allometric models}
\usage{
AllometricModel(
  response_unit,
  covariate_units,
  predict_fn,
  descriptors = list()
)
}
\arguments{
\item{response_unit}{\itemize{
\item A list containing one element with the name of the
respone variable as the key and the units of the response variable as class
\code{units::units}
}}

\item{covariate_units}{\itemize{
\item A list containing all covariates used in the model
with the name of the covariate as the key and the units of the covariate as
class \code{units::units}
}}

\item{predict_fn}{\itemize{
\item A function that takes all covariates named in
\code{covariate_units} as arguments.
}}
}
\description{
This class is primarily used as a parent class for other model
implementations.
}
