% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_checks.R
\name{check_args_in_predict_fn}
\alias{check_args_in_predict_fn}
\title{Check if the arguments of the predict_fn are all in the predict_fn body.}
\usage{
check_args_in_predict_fn(object)
}
\description{
Check if the arguments of the predict_fn are all in the predict_fn body.
}
\keyword{internal}
