% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{check_ids_equal}
\alias{check_ids_equal}
\title{Check for equivalence of ID slots in two models}
\usage{
check_ids_equal(mod1, mod2)
}
\arguments{
\item{mod1}{A model object}

\item{mod2}{A model object}
}
\value{
TRUE if equal, FALSE if not
}
\description{
Two models are considered to have equal IDs if (1) they both do not have ID
slots, which can occur especially in testing, or (2) both models have ID
slots and the value of the slots are equal. They are not equal otherwise.
}
\keyword{internal}
