% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{check_list_equal}
\alias{check_list_equal}
\title{Check for equivalence of two lists}
\usage{
check_list_equal(list1, list2)
}
\description{
Two lists are considered equal if all values are the same and all names are
the same. Order is not considered.
}
\keyword{internal}
