% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{descriptors}
\alias{descriptors}
\title{Get the descriptors of a model}
\usage{
descriptors(object)
}
\arguments{
\item{object}{The allometric model or model set object}
}
\value{
A tibble:tbl_df of descriptors
}
\description{
The model descriptors describe the context of an allometric model as it is
situated within a publication, and contain information like family, genus,
species, geographic region, etc. This function returns this information for
a given model.
}
\keyword{internal}
