% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{download_models}
\alias{download_models}
\title{Download allometric models}
\usage{
download_models(verbose)
}
\description{
This function downloads allometric models from GitHub into the local package
directory. Any existing models are removed before downloading.
}
\keyword{internal}
