% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffectsModel.R
\name{FixedEffectsModel}
\alias{FixedEffectsModel}
\title{Instantiate a fixed effects model.}
\usage{
FixedEffectsModel(
  response_unit,
  covariate_units,
  predict_fn,
  parameters,
  descriptors = list()
)
}
\description{
This class is essentially a wrapper around \code{ParametricModel} but clarifies
the intent between \code{FixedEffectsModel} and \code{MixedEffectsModel}.
}
