% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixedEffectsSet.R
\name{FixedEffectsSet}
\alias{FixedEffectsSet}
\title{Create a set of fixed effects models}
\usage{
FixedEffectsSet(
  response_unit,
  covariate_units,
  parameter_names,
  predict_fn,
  model_specifications,
  descriptors = list(),
  covariate_definitions = list()
)
}
\arguments{
\item{response_unit}{A named list containing one element, with a name representing the response
variable and a value representing the units of the response variable
using the \code{units::as_units} function.}

\item{covariate_units}{A named list containing the covariate specifications, with names
representing the covariate name and the values representing the units of
the coavariate using the \code{units::as_units} function}

\item{parameter_names}{A character vector naming the columns in \code{model_specifications} that
represent the parameters}

\item{predict_fn}{A function that takes the covariate names as arguments and returns a
prediction of the response variable. This function should be vectorized.}

\item{model_specifications}{A dataframe such that each row of the dataframe provides model-level
descriptors and parameter estimates for that model. Models must be
uniquely identifiable using the descriptors}

\item{descriptors}{An optional named list of descriptors that describe the context of the
allometric model}

\item{covariate_definitions}{An optional named list of custom covariate definitions that will supersede
the definitions given by the variable naming system.}
}
\value{
A set of fixed effects models
}
\description{
A \code{FixedEffectsSet} represents a group of fixed-effects models that all have
the same functional structure. Fitting a large family of models (e.g., for
many different species) using the same functional structure is a common
pattern in allometric studies, and \code{FixedEffectsSet} facilitates the
installation of these groups of models by allowing the user to specify the
parameter estimates and descriptions in a dataframe or spreadsheet.
}
