% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_processing.R
\name{ingest_models}
\alias{ingest_models}
\title{Ingest a set of models by running the publication files}
\usage{
ingest_models(verbose, pub_path = NULL, params_path = NULL)
}
\arguments{
\item{verbose}{If TRUE, print verbose messages}

\item{pub_path}{A path to a directory containing publication files}

\item{params_path}{A path to a directory containing parameter files}
}
\description{
Ingest a set of models by running the publication files
}
