% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_models}
\alias{load_models}
\title{Load the locally installed table of allometric models}
\usage{
load_models()
}
\value{
A model_tbl containing the locally installed models.
}
\description{
This function loads all locally installed allometric models if they are
downloaded and installed, if not run the \code{install_models} function. The
result is of class \code{model_tbl}, which behaves very much like a
\code{tibble::tbl_df} or a \code{data.frame}.
}
\details{
Printing the \code{head} of \code{allometric_models}, we can see the structure of the
data

\if{html}{\out{<div class="sourceCode r">}}\preformatted{allometric_models <- load_models()
#> Warning: 74 publication(s) lack required BibTeX fields for their declared bibtype; citations were built with bibtype 'misc': barnes_1962, barrett_1978, barrett_2006, bell_1981, bluhm_2007, ...
head(allometric_models)
#> # A tibble: 6 x 12
#>   id       spec_index model_name model_type  pub_id       pub_year family_name covt_name taxa              region component model     
#>   <chr>         <int> <chr>      <chr>       <chr>           <dbl> <list>      <list>    <list> <list<character>> <chr>     <list>    
#> 1 1de6301a          0 hstix50    site index  barnes_1962      1962 <chr [1]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 2 0f54ec1a          0 hstix100   site index  barrett_1978     1978 <chr [1]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 3 fdc26756          0 hst        stem height barrett_2006     2006 <chr [1]>   <chr [1]> <Taxa>                   <NA>      <FxdEffcM>
#> 4 fdc26756          1 hst        stem height barrett_2006     2006 <chr [1]>   <chr [1]> <Taxa>                   <NA>      <FxdEffcM>
#> 5 fdc26756          2 hst        stem height barrett_2006     2006 <chr [1]>   <chr [1]> <Taxa>                   <NA>      <FxdEffcM>
#> 6 fdc26756          3 hst        stem height barrett_2006     2006 <chr [1]>   <chr [1]> <Taxa>                   <NA>      <FxdEffcM>
}\if{html}{\out{</div>}}

The columns are:
\itemize{
\item \code{id} - A unique ID for the model. This is the 8-character content hash
assigned by the v4 model compilation pipeline.
\item \code{spec_index} - The index of the specification within its model set (0 for
single models). Together with \code{id} this uniquely identifies a model.
\item \code{model_name} - The name of the model or model set.
\item \code{model_type} - The type of model (e.g., stem volume, site index, etc.)
\item \code{pub_id} - A unique ID representing the publication.
\item \code{pub_year} - The publication year.
\item \code{family_name} - The names of the contributing authors.
\item \code{covt_name} - The names of the covariates used in the model.
\item \code{taxa} - The taxonomic specification of the trees that are modeled.
\item \code{region} - The region or regions (e.g., state, province, etc.) from which
the model data is from.
\item \code{component} - The tree component modeled (e.g., stem, branch).
\item \code{model} - The model object itself.
}

Models can be searched by their attributes. Note that some of the columns
are \code{list} columns, which contain lists as their elements. Filtering on
data in these columns requires the use of \code{purrr::map_lgl} which is used to
determine truthiness of expressions for each element in a \code{list} column.
While this may seem complicated, we believe the nested data structures are
more descriptive and concise for storing the models, and users will quickly
find that searching models in this way can be very powerful.
}
\section{Finding Contributing Authors}{
Using \code{purr::map_lgl} to filter the \code{family_name} column, we are able to
find publications that contain specific authors of interst. For example, we
may want models only authored by \code{"Hann"}. This is elementary to do in
\code{allometric}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{hann_models <- dplyr::filter(
 allometric_models,
 purrr::map_lgl(family_name, ~ 'Hann' \%in\% .)
)

head(hann_models)
#> # A tibble: 6 x 12
#>   id       spec_index model_name model_type  pub_id    pub_year family_name covt_name taxa              region component model     
#>   <chr>         <int> <chr>      <chr>       <chr>        <dbl> <list>      <list>    <list> <list<character>> <chr>     <list>    
#> 1 713c7427          0 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 2 713c7427          1 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 3 713c7427          2 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 4 713c7427          3 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 5 713c7427          4 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 6 713c7427          5 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
nrow(hann_models)
#> [1] 76
}\if{html}{\out{</div>}}

Picking apart the above code block, we see that we are using the standard
\code{dplyr::filter} function on the \code{allometric_models} dataframe. The second
argument is a call using \code{purrr:map_lgl}, which will map over each list
(contained as elements in the \code{family_names} column). The second argument to
this function, \code{~ 'Hann' \%in\% .} is itself a function that checks if \code{'Hann'}
is in the current list. Imagine we are marching down each row of
\code{allometric_models}, \code{.} represents the element of \code{family_names} we are
considering, which is itself a list of author names.
}

\section{Finding First Authors}{
Maybe we are only interested in models where \code{'Hann'} is the first author.
Using a simple modification we can easily do this.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{hann_first_author_models <- dplyr::filter(
  allometric_models,
  purrr::map_lgl(family_name, ~ 'Hann' == .[[1]])
)

head(hann_first_author_models)
#> # A tibble: 6 x 12
#>   id       spec_index model_name model_type  pub_id    pub_year family_name covt_name taxa              region component model     
#>   <chr>         <int> <chr>      <chr>       <chr>        <dbl> <list>      <list>    <list> <list<character>> <chr>     <list>    
#> 1 713c7427          0 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 2 713c7427          1 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 3 713c7427          2 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 4 713c7427          3 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 5 713c7427          4 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 6 713c7427          5 vsia       stem volume hann_1978     1978 <chr [2]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
nrow(hann_first_author_models)
#> [1] 42
}\if{html}{\out{</div>}}

We can see that \code{'Hann'} is the first author for
42 models in this package.
}

\section{Finding Models for a Given Species}{
One of the most common things people need is a model for a particular
species. For this, we must interact with the \code{taxa} column. For example,
to find models for the Pinus genus we can use

\if{html}{\out{<div class="sourceCode r">}}\preformatted{pinus_models <- dplyr::filter(
 allometric_models,
 purrr::map_lgl(taxa, ~ "Pinus" \%in\% .)
)

head(pinus_models)
#> # A tibble: 6 x 12
#>   id       spec_index model_name model_type  pub_id       pub_year family_name covt_name taxa              region component model     
#>   <chr>         <int> <chr>      <chr>       <chr>           <dbl> <list>      <list>    <list> <list<character>> <chr>     <list>    
#> 1 0f54ec1a          0 hstix100   site index  barrett_1978     1978 <chr [1]>   <chr [2]> <Taxa>                   <NA>      <FxdEffcM>
#> 2 fdc26756         17 hst        stem height barrett_2006     2006 <chr [1]>   <chr [1]> <Taxa>                   <NA>      <FxdEffcM>
#> 3 fdc26756         18 hst        stem height barrett_2006     2006 <chr [1]>   <chr [1]> <Taxa>                   <NA>      <FxdEffcM>
#> 4 fdc26756         19 hst        stem height barrett_2006     2006 <chr [1]>   <chr [1]> <Taxa>                   <NA>      <FxdEffcM>
#> 5 fdc26756         20 hst        stem height barrett_2006     2006 <chr [1]>   <chr [1]> <Taxa>                   <NA>      <FxdEffcM>
#> 6 fdc26756         21 hst        stem height barrett_2006     2006 <chr [1]>   <chr [1]> <Taxa>                   <NA>      <FxdEffcM>
nrow(pinus_models)
#> [1] 324
}\if{html}{\out{</div>}}

Users can also search with a specific taxon, which allows a full
specification from family to species. For example, if we want models that
apply to Ponderosa pine, first declare the necessary taxon, then use it to
filter as before

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ponderosa_taxon <- Taxon(
 family = "Pinaceae", genus = "Pinus", species = "ponderosa"
)

ponderosa_models <- dplyr::filter(
 allometric_models,
 purrr::map_lgl(taxa, ~ ponderosa_taxon \%in\% .)
)

nrow(ponderosa_models)
#> [1] 57
}\if{html}{\out{</div>}}
}

\section{Finding a Model with Specific Data Requirements}{
We can even check for models that contain certain types of data requirements.
For example, the following block finds diameter-height models, specifically
models that use diameter outside bark at breast height as the \emph{only}
covariate. The utility here is obvious, since many inventories are vastly
limited by their available tree measurements.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{dia_ht_models <- dplyr::filter(
    allometric_models,
    model_type == 'stem height',
    purrr::map_lgl(covt_name, ~ length(.)==1 & .[[1]] == 'dsob'),
)

nrow(dia_ht_models)
#> [1] 284
}\if{html}{\out{</div>}}

Breaking this down, we have the first condition \code{model_type=='stem_height'}
selecting only models concerned with stem heights as a response variable. The
second line maps over each element of the \code{covt_name} column, which is a
character vector. The \code{.} represents a given character vector for that row.
First, we ensure that the vector is only one element in size using
\code{length(.)==1}, then we ensure that the first (and only) element of this
vector is equal to \code{'dsob'}, (diameter outside bark at breast height). In
this case, 284 are available in the package.
}

\section{Finding a Model for a Region}{
By now the user should be sensing a pattern. We can apply the exact same
logic as the \emph{Finding Contributing Authors} section to find all models
developed using data from \code{US-CO}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{us_co_models <- dplyr::filter(
    allometric_models,
    purrr::map_lgl(region, ~ "US-CO" \%in\% .),
)

nrow(us_co_models)
#> [1] 6
}\if{html}{\out{</div>}}

We can see that 6 allometric models are defined for the
state of Colorado, US.
}

