% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_parameter_frame}
\alias{load_parameter_frame}
\title{Load a parameter frame from the models/parameters directory}
\usage{
load_parameter_frame(name)
}
\arguments{
\item{name}{The name of the file, excluding the extension}
}
\value{
A tibble::tbl_df of the parameter data.
}
\description{
This is a convenience that allows a user to easily load parameter files from
the models/parameters directory. It is typically used when constructing
the \code{model_specifications} argument for \code{ModelSet}.
}
