% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedEffectsModel.R
\name{MixedEffectsModel}
\alias{MixedEffectsModel}
\title{Instantiate a fixed effects model.}
\usage{
MixedEffectsModel(
  response_unit,
  covariate_units,
  predict_ranef,
  predict_fn,
  parameters,
  fixed_only = FALSE,
  descriptors = list()
)
}
\description{
This class is essentially a wrapper around \code{ParametricModel} but clarifies
the intent between \code{FixedEffectsModel} and \code{MixedEffectsModel}.
}
