% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixedEffectsSet.R
\name{MixedEffectsSet}
\alias{MixedEffectsSet}
\title{Create a set of mixed effects models}
\usage{
MixedEffectsSet(
  response_unit,
  covariate_units,
  parameter_names,
  predict_fn,
  model_specifications,
  predict_ranef,
  fixed_only = FALSE,
  descriptors = list(),
  covariate_definitions = list()
)
}
\arguments{
\item{response_unit}{A named list containing one element, with a name representing the response
variable and a value representing the units of the response variable
using the \code{units::as_units} function.}

\item{covariate_units}{A named list containing the covariate specifications, with names
representing the covariate name and the values representing the units of
the coavariate using the \code{units::as_units} function.}

\item{parameter_names}{A character vector naming the columns in \code{model_specifications} that
represent the parameters.}

\item{model_specifications}{A dataframe such that each row of the dataframe provides model-level
descriptors and parameter estimates for that model. Models must be
uniquely identifiable using the descriptors.}

\item{predict_ranef}{A function that predicts the random effects, takes any named covariates in
\code{covariate_units} as arguments.}

\item{fixed_only}{A boolean value indicating if the model produces predictions using only
fixed effects. This is useful when publications do not provide sufficient
information to predict the random effects.}

\item{descriptors}{An optional list of descriptors that are specified at the model-level.}

\item{covariate_definitions}{A named list of covariate definitions that
are used instead of the descriptions given by the variable naming system.}
}
\description{
A \code{MixedEffectsSet} represents a group of fixed-effects models that all have
the same functional structure. Fitting a large family of models (e.g., for
many different species) using the same functional structure is a common
pattern in allometric studies, and \code{MixedEffectsSet} facilitates the
installation of these groups of models by allowing the user to specify the
parameter estimates and descriptions in a dataframe or spreadsheet.
}
\details{
Because mixed-effects models already accommodate a grouping structure,
\code{MixedEffectsSet} tends to be a much rarer occurrence than \code{FixedEffectsSet}
and \code{MixedEffectsModel}.
}
