% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{model_call}
\alias{model_call}
\title{Get the function call for a model}
\usage{
model_call(model)
}
\arguments{
\item{model}{The allometric model for which a function call will be
retrieved.}
}
\description{
This function allows a user to see the structure of the function call for
a given model in an easy-to-read format.
}
