% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParametricModel.R
\name{ParametricModel}
\alias{ParametricModel}
\title{Base class for all parametric models.}
\usage{
ParametricModel(
  response_unit,
  covariate_units,
  predict_fn,
  parameters,
  descriptors = list(),
  covariate_definitions = list()
)
}
\description{
This is a base class used for \code{FixedEffectsModel} and \code{MixedEffectsModel}
}
\keyword{internal}
