\name{poudel_2019}
\alias{poudel_2019}
\title{Poudel, Temesgen, Radtke et al. (2019)}
\description{Allometric models from Poudel, Temesgen, Radtke et al. (2019)}
\section{Stem Volume Models}{\code{vsa = exp(a + b * log(dsob) + c * log(hst))}
\itemize{
\item{\code{vsa [m^3]}}{ - volume of the entire stem, including top and stump}
\item{\code{dsob [cm]}}{ - diameter of the stem, outside bark at breast height}
\item{\code{hst [m]}}{ - total height of the stem}
}
\bold{Model Parameters}
\preformatted{   family       genus         species           a     b     c
 1 Pinaceae     Abies         concolor     -10.1   1.60 1.34 
 2 Pinaceae     Abies         grandis       -9.18  1.55 1.17 
 3 Pinaceae     Abies         lasiocarpa    -9.85  1.70 1.15 
 4 Pinaceae     Abies         magnifica    -10.3   1.69 1.31 
 5 Cupressaceae Chamaecyparis nootkatensis  -9.60  1.84 0.95 
 6 Pinaceae     Larix         occidentalis -11.3   1.93 1.33 
 7 Cupressaceae Calocedrus    decurrens     -9.64  1.66 1.08 
 8 Pinaceae     Picea         <NA>         -10.4   1.82 1.19 
 9 Pinaceae     Picea         engelmannii  -10.3   1.69 1.32 
10 Pinaceae     Picea         glauca       -10.6   2.12 0.894
11 Pinaceae     Picea         sitchensis   -10.1   1.86 1.08 
12 Pinaceae     Pinus         contorta     -10.0   1.86 1.06 
13 Pinaceae     Pinus         flexilis     -11.4   2.14 1.20 
14 Pinaceae     Pinus         jeffreyi     -10.5   2.03 0.941
15 Pinaceae     Pinus         lambertiana  -10.7   2.44 0.489
16 Pinaceae     Pinus         monticola     -9.60  1.40 1.41 
17 Pinaceae     Pinus         ponderosa    -10.6   2.11 0.912
18 Pinaceae     Pseudotsuga   menziesii     -9.55  1.74 1.02 
19 Cupressaceae Sequoia       sempervirens -10.8   2.05 0.947
20 Cupressaceae Thuja         plicata       -9.55  2.04 0.66 
21 Pinaceae     Tsuga         <NA>          -9.47  1.83 0.936
22 Pinaceae     Tsuga         heterophylla  -9.98  1.96 0.925
23 Pinaceae     Tsuga         mertensiana  -11.0   1.88 1.30 
24 Betulaceae   Alnus         rubra        -10.3   2.05 0.944
25 Salicaceae   Populus       tremuloides  -10.7   1.90 1.24 }
\out{<hr>}}
