% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/FixedEffectsModel.R,
%   R/MixedEffectsModel.R
\name{predict}
\alias{predict}
\alias{predict,FixedEffectsModel-method}
\alias{predict,MixedEffectsModel-method}
\title{Predict with an allometric model}
\usage{
predict(mod, ...)

\S4method{predict}{FixedEffectsModel}(mod, ..., output_units = NULL)

\S4method{predict}{MixedEffectsModel}(mod, ..., newdata = NULL, output_units = NULL)
}
\arguments{
\item{mod}{The allometric model used for prediction}

\item{...}{Additional arguments passed to the \code{predict_fn} of the input model}

\item{output_units}{Optionally specify the output units of the model as a string, e.g., "ft^3"}

\item{newdata}{A dataframe containing columns that match the names of the arguments given
to \code{predict_ranef}. The values of this data represents information from a
new group of observations for which predictions are desired (e.g., a new
stand or plot).}
}
\description{
Predict with an allometric model
}
