% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_tbl.R
\name{predict_allo}
\alias{predict_allo}
\title{Predict allometric attributes using a column of allometric models}
\usage{
predict_allo(model_list, ...)
}
\arguments{
\item{model_list}{A list of models, usually obtained from \code{allometric_models}
\code{model} column.}

\item{...}{Additional arguments passed to each model's \code{predict_fn}}
}
\value{
A vector of predictions
}
\description{
Predict allometric attributes using a column of allometric models
}
