% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{set_params_path}
\alias{set_params_path}
\title{Set the parameter search path}
\usage{
set_params_path(params_path)
}
\arguments{
\item{params_path}{The file path containing parameter files}
}
\description{
The parameter search path is where \code{allometric} looks for parameter frames.
By default, the package searches the local installation, however it is
useful when testing publication files to search a local directory, which
can be set here.
}
